requirements.txt
Flask
Flask-Cors
google-auth-oauthlib
google-api-python-client
google-generativeai
telethon
python-dotenv
beautifulsoup4
scikit-learn
gunicorn